# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ayoitson API
  version: 1.0.0
  description: |
    The Ayoitson REST API. Phase 4 redesign — uniform POST/PUT/DELETE
    semantics across every resource, structured error envelope on every
    response, X-API-Key auth on every endpoint except `GET /api/health` and
    `POST /api/auth/setup`. See `docs/SDLC.md` and the vault `[[API Redesign]]`
    note for the design memo.
  contact:
    name: SALM LLC / Ayoitson
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - apiKey: []
...
paths:
...
  /api/plex-servers:
    get:
      tags: [plex-servers]
      summary: List Plex servers
      description: Auth-required because the response discloses server identity / URL even with tokens redacted.
      responses:
        '200':
          description: Plex server list (tokens redacted)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlexServerPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [plex-servers]
      summary: Create Plex server
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlexServerCreate' }
      responses:
        '201':
          description: Plex server created
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: boolean }
                  name: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/plex-servers/{name}:
    parameters:
      - $ref: '#/components/parameters/PlexServerName'
    put:
      tags: [plex-servers]
      summary: Update Plex server
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlexServerCreate' }
      responses:
        '200':
          description: Plex server updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlexServerUpdateResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [plex-servers]
      summary: Delete Plex server
      responses:
        '200':
          description: Plex server deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteAck' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/plex-servers/{name}/status-check:
    parameters:
      - $ref: '#/components/parameters/PlexServerName'
    post:
      tags: [plex-servers]
      summary: Probe a configured Plex server
      responses:
        '200':
          description: Probe result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlexStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/plex-servers/foreign-status-check:
    post:
      tags: [plex-servers]
      summary: Probe an arbitrary Plex server (body)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlexServerCreate' }
      responses:
        '200':
          description: Probe result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlexStatus' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/plex-servers/status:
    post:
      tags: [plex-servers]
      summary: Legacy Plex status helper
      deprecated: true
      description: Legacy compatibility route. Prefer `POST /api/plex-servers/{name}/status-check`.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200':
          description: Probe result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlexStatus' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/settings/plex:
    get:
      tags: [settings]
      summary: Read Plex settings
      responses:
        '200': { description: Plex settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [settings]
      summary: Update Plex settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Updated Plex settings, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Issued by the first-run setup flow (`scripts/first-run.js` or
        `POST /api/auth/setup`). Format: `ayo_<base64url>`. Verified server-
        side via argon2id against the `api_keys` SQLite table.
  parameters:
...
    PlexServerName:
      name: name
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - PAYLOAD_TOO_LARGE
            - RATE_LIMITED
            - INTERNAL
            - UPSTREAM_ERROR
            - SERVICE_UNAVAILABLE
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    PlexServerPublic:
      type: object
      required: [name]
      properties:
        name: { type: string }
        uri: { type: string }
        index: { type: integer }
        arGuide: { type: boolean }
        arChannels: { type: boolean }
    PlexServerCreate:
      type: object
      required: [name, uri, accessToken]
      properties:
        name: { type: string }
        uri: { type: string }
        accessToken: { type: string }
        arGuide: { type: boolean }
        arChannels: { type: boolean }
    PlexServerUpdateResult:
      type: object
      properties:
        updated: { type: boolean }
        name: { type: string }
        modifiedPrograms: { type: integer }
        destroyedPrograms: { type: integer }
    PlexStatus:
      type: object
      properties:
        status: { type: integer }
...
    DeleteAck:
      type: object
      properties:
        deleted: { type: boolean }
        id: { type: string }
        number: { type: integer }
        name: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid X-API-Key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Request body or parameter is invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...